Fixes manifest circular dependency (#6860)
