+++ src/OrchardCore.Modules/OrchardCore.Contents/Liquid/ContentItemFilter.cs
@@ -23,7 +23,6 @@ namespace OrchardCore.Contents.Liquid
             if (input.Type == FluidValues.Array)
             {
                 // List of content item ids
-
                 var contentItemIds = input.Enumerate().Select(x => x.ToStringValue()).ToArray();
 
                 return FluidValue.Create(await contentManager.GetAsync(contentItemIds));
