+++ src/OrchardCore.Modules/OrchardCore.Liquid/Manifest.cs
@@ -6,6 +6,5 @@ using OrchardCore.Modules.Manifest;
     Website = ManifestConstants.OrchardCoreWebsite,
     Version = ManifestConstants.OrchardCoreVersion,
     Description = "The liquid module enables content items to have liquid syntax.",
-    Dependencies = new[] { "OrchardCore.Contents", "OrchardCore.Shortcodes" },
     Category = "Content Management"
 )]
