+++ src/OrchardCore.Themes/TheAdmin/Recipes/headless.recipe.json
@@ -48,7 +48,6 @@
         "OrchardCore.Layers",
         "OrchardCore.Lucene",
         "OrchardCore.Lists",
-        "OrchardCore.Liquid",
         "OrchardCore.Markdown",
         "OrchardCore.Media",
         "OrchardCore.Menu",
