+++ src/OrchardCore.Themes/TheAgencyTheme/Recipes/agency.recipe.json
@@ -50,7 +50,6 @@
         "OrchardCore.Flows",
         "OrchardCore.Layers",
         "OrchardCore.Lists",
-        "OrchardCore.Liquid",
         "OrchardCore.Markdown",
         "OrchardCore.Media",
         "OrchardCore.Menu",
