+++ src/OrchardCore.Themes/TheBlogTheme/Recipes/blog.recipe.json
@@ -61,7 +61,6 @@
         "OrchardCore.Layers",
         "OrchardCore.Lucene",
         "OrchardCore.Lists",
-        "OrchardCore.Liquid",
         "OrchardCore.Markdown",
         "OrchardCore.Media",
         "OrchardCore.Menu",
