+++ src/OrchardCore.Themes/TheComingSoonTheme/Recipes/comingsoon.recipe.json
@@ -47,7 +47,6 @@
         "OrchardCore.HomeRoute",
         "OrchardCore.Html",
 
-        "OrchardCore.Liquid",
         "OrchardCore.Localization",
 
         "OrchardCore.Media",
@@ -56,7 +55,6 @@
         "OrchardCore.Recipes",
         "OrchardCore.Resources",
         "OrchardCore.Roles",
-
 
         "OrchardCore.Scripting",
         "OrchardCore.Settings",
